+++ src/QueryBuilder/QueryBuilder.php
@@ -35,4 +35,19 @@ class QueryBuilder extends AbstractQueryBuilder
     {
         return parent::setArgument($argumentName, $argumentValue);
     }
+
+    /**
+     * Changing method visibility to public
+     *
+     * @param string $name
+     * @param string $type
+     * @param bool   $isRequired
+     * @param null   $defaultValue
+     *
+     * @return AbstractQueryBuilder|QueryBuilder
+     */
+    public function setVariable(string $name, string $type, bool $isRequired = false, $defaultValue = null)
+    {
+        return parent::setVariable($name, $type, $isRequired, $defaultValue);
+    }
 }