+++ tests/ClientTest.php
@@ -57,6 +57,7 @@ class ClientTest extends TestCase
 
         $mockHandler->append(new Response(200));
         $mockHandler->append(new Response(200));
+        $mockHandler->append(new Response(200));
 
         $client = new MockClient('', $handler);
         $client->runRawQuery('query_string');
@@ -64,17 +65,24 @@ class ClientTest extends TestCase
         $client = new MockClient('', $handler, ['Authorization' => 'Basic xyz']);
         $client->runRawQuery('query_string');
 
+        $client = new MockClient('', $handler);
+        $client->runRawQuery('query_string',  false, ['name' => 'val']);
+
         /** @var Request $firstRequest */
         $firstRequest = $container[0]['request'];
-        $this->assertEquals('{"query":"query_string"}', $firstRequest->getBody()->getContents());
+        $this->assertEquals('{"query":"query_string","variables":{}}', $firstRequest->getBody()->getContents());
 
-        /** @var Request $secondRequest */
-        $secondRequest = $container[1]['request'];
-        $this->assertNotNull($secondRequest->getHeader('Authorization'));
+        /** @var Request $thirdRequest */
+        $thirdRequest = $container[1]['request'];
+        $this->assertNotNull($thirdRequest->getHeader('Authorization'));
         $this->assertEquals(
             ['Basic xyz'],
-            $secondRequest->getHeader('Authorization')
+            $thirdRequest->getHeader('Authorization')
         );
+
+        /** @var Request $secondRequest */
+        $secondRequest = $container[2]['request'];
+        $this->assertEquals('{"query":"query_string","variables":{"name":"val"}}', $secondRequest->getBody()->getContents());
     }
 
     /**